Move resource base file to resources directory
