+++ src/resources/_resource-base.js
