+++ src/resources/listings.js
@@ -1,8 +1,8 @@
 // For now, we are just wrapping the methods that are already in
 // contractService and ipfsService.
-import ResourceBase from"../ResourceBase"
+import ResourceBase from "./_resource-base"
 
-class Listings extends ResourceBase{
+class Listings extends ResourceBase {
   constructor({ contractService, ipfsService }) {
     super({ contractService, ipfsService })
     this.contractDefinition = this.contractService.listingContract
